% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util_Func.R
\name{CensDataGen}
\alias{CensDataGen}
\title{Generate the censored data for simulation}
\usage{
CensDataGen(True_Data, dl)
}
\arguments{
\item{True_Data}{example data generated by TrueDataGen}

\item{dl}{the censored limit}
}
\value{
a censored example data
}
\description{
Generate the censored data for simulation
}
