% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util_Func.R
\name{eval_density}
\alias{eval_density}
\title{Evaluate the density of subject}
\usage{
eval_density(y, c, m, v)
}
\arguments{
\item{y}{a p-dimension vectors of continous values: observed manifest vars}

\item{c}{a p-dimension vectors of 0,-1,1 indicating censoring directions}

\item{m}{a p-dimension vectors of continous values: means estimated by the regression}

\item{v}{a pxp varcov matrix}
}
\value{
a positive real value of the density
}
\description{
Evaluate the density of subject
}
