% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util_Func.R
\name{eval_r}
\alias{eval_r}
\title{Evaluate r in multivariate EM step.}
\usage{
eval_r(y, c, m, v)
}
\arguments{
\item{y}{a p-dimension vectors of continous values: observed manifest vars.}

\item{c}{a p-dimension vectors of 0,-1,1 indicating censoring directions.}

\item{m}{a p-dimension vectors of continous values: means.}

\item{v}{a pxp varcov matrix.}
}
\value{
a pxp cond_varcov matrix subjected to 0s for the obs features.
}
\description{
Evaluate r in multivariate EM step.
}
