% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util_Func.R
\name{eval_ystar}
\alias{eval_ystar}
\title{Evaluate the Y-star in EM}
\usage{
eval_ystar(y, c, m, v)
}
\arguments{
\item{y}{a p-dimension vectors of continous values: observed manifest vars}

\item{c}{a p-dimension vectors of 0,-1,1 indicating censoring directions}

\item{m}{a p-dimension vectors of continous values: means}

\item{v}{a pxp varcov matrix}
}
\value{
a p-dimension vector of obs_y or conditional means of cen_y
}
\description{
Evaluate the Y-star in EM
}
