% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixCenMVReg_EM.R
\name{MixCenMVReg_EM}
\alias{MixCenMVReg_EM}
\title{Censored Mixture Multivariate Regression EM}
\usage{
MixCenMVReg_EM(
  Y,
  C,
  X,
  G = 2,
  Max.iter = 1000,
  pie_hat = NA,
  beta_hat = NA,
  sigma_hat = NA,
  diff.tol = 0.001,
  print = TRUE,
  init_class = NA,
  calc_cov = FALSE
)
}
\arguments{
\item{Max.iter}{the maximum iteration numbers}

\item{pie_hat}{a vector of initialized Mixing porportions}

\item{beta_hat}{a list of initialized Beta matrices}

\item{sigma_hat}{a list of initialized Sigma matrices}

\item{diff.tol}{the judgement standard of convergence, default value is 1e-3.}

\item{calc_cov}{}
}
\value{
Function outputs a list including the following:
\describe{
      \item{\code{Iterations}}{The number of iterations of EM in the process.}
      \item{\code{Converged}}{Whether it is converged or not.}
      \item{\code{LogLik}}{the LogLikelihood for the model fitting.}
      \item{\code{AIC}}{Akaike's Information Criterion}
      \item{\code{BIC}}{Bayesian Information Criterion}
      \item{\code{ICL}}{Integrated Complete-data Likelihood Criterion}
      \item{\code{Pie}}{A vector of estimated Mixing porportions}
      \item{\code{Beta}}{A list of estimated Coefficient matrices}
      \item{\code{Sigma}}{a list of estimated initialized Covariance matrices}
      \item{\code{Posterior}}{Posterior probability for each observation in each class}
      \item{\code{Class}}{The predicted outcome of each observation}
      \item{\code{obs_Info}}{}
}
}
\description{
Censored Mixture Multivariate Regression EM
}
