% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixCenUVReg_EM.R
\name{MixCenUVReg_EM}
\alias{MixCenUVReg_EM}
\title{Censored Mixture Univariate Regression EM}
\usage{
MixCenUVReg_EM(
  Y,
  C,
  X,
  G = 2,
  Max.iter = 1000,
  pie_hat = NA,
  beta_hat = NA,
  sigma_hat = NA,
  diff.tol = 0.001,
  print = TRUE,
  init_class = NA,
  calc_cov = FALSE
)
}
\arguments{
\item{Y}{the response variables matrix, N x p.}

\item{C}{censored ID matrix, N x p, -1 means left, 0 means observed and 1 means right censored.}

\item{X}{covariates matrix, N x d, with d-1 covariates.}

\item{G}{the given number of groups.}

\item{Max.iter}{the maximum iteration numbers.}

\item{pie_hat}{a vector of initialized Mixing porportions.}

\item{beta_hat}{a list of initialized Beta matrices, each matrix with d x p, with p response variables and d-1 covariates.}

\item{sigma_hat}{a list of initialized Sigma matrices, each matrix with p x p.}

\item{diff.tol}{the judgement standard of convergence, default value is 1e-3.}

\item{print}{whether to print the outcome.}

\item{init_class}{initial partition, which needs to be factor format.}

\item{calc_cov}{whether to calculate the covariance matrix.}
}
\value{
Function outputs a list including the following:
\describe{
      \item{\code{Iterations}}{The number of iterations of EM in the process.}
      \item{\code{Converged}}{Whether it is converged or not.}
      \item{\code{LogLik}}{the LogLikelihood for the model fitting.}
      \item{\code{AIC}}{Akaike's Information Criterion}
      \item{\code{BIC}}{Bayesian Information Criterion}
      \item{\code{ICL}}{Integrated Complete-data Likelihood Criterion}
      \item{\code{Pie}}{A vector of estimated Mixing porportions}
      \item{\code{Beta}}{A list of estimated Coefficient matrices}
      \item{\code{Sigma}}{a list of estimated initialized Covariance matrices}
      \item{\code{Posterior}}{Posterior probability for each observation in each class}
      \item{\code{Class}}{The predicted outcome of each observation}
      \item{\code{obs_Info}}{false if choose calc_cov=false}
      \item{\code{Cov}}{cov matrix}
}
}
\description{
Censored Mixture Univariate Regression EM
}
