% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util_Func.R
\name{TrueDataGen}
\alias{TrueDataGen}
\title{Generate data for example}
\usage{
TrueDataGen(Replicate = 30, n = 1000, pie, beta, sigma)
}
\arguments{
\item{Replicate}{replicate time}

\item{n}{subject numbers}

\item{pie}{a list of pie value}

\item{beta}{a list of beta matrices}

\item{sigma}{a list of sigma matrices}
}
\value{
a list of data
}
\description{
Generate data for example
}
\examples{
set.seed(22)
BETA=list(matrix(c(2,0,20,-2),nrow=2,ncol=2),matrix(c(3,1,25,-3),nrow=2,ncol=2),matrix(c(3.5,2,30,-5),nrow=2,ncol=2))
SIGMA=list(matrix(c(1,0.1,0.1,1),nrow=2,ncol=2),matrix(c(2,0.2,0.2,0.5),nrow=2,ncol=2),matrix(c(0.5,0.3,0.3,2),nrow=2,ncol=2))
PIE=c(.1,.7,.2)
example_true=TrueDataGen(1,1000,PIE,BETA,SIGMA)

}
