% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util_Func.R
\name{TrueDataGen}
\alias{TrueDataGen}
\title{True data generation}
\usage{
TrueDataGen(Replicate = 30, n = 1000, pie, beta, sigma)
}
\arguments{
\item{Replicate}{the replicate number}

\item{n}{the number of observations}

\item{pie}{a bector of Mixing proportions}

\item{beta}{a list of Beta matrices}

\item{sigma}{a list of Sigma matrices}
}
\value{
a list of true data for simulation
}
\description{
True data generation
}
